Update 4.7 swap2num.c

+++ 4.7 swap2num.c
@@ -2,9 +2,9 @@
 #include<conio.h>
 void main()
 {
-    float a=0,b=0,c=0;
+    int a=0,b=0,c=0;
     printf("Enter 2 number");
-    scanf("%f%f",&a,&b);
+    scanf("%d%d",&a,&b);
     c=a;
     a=b;
     b=c;
